requirements.txt
seaborn==0.11.1
matplotlib==3.3.4
tabula_py==2.3.0
pandas==1.2.3
numpy==1.19.5
streamlit==0.86.0
camelot-py[base]
Pillow==8.3.2
tabula-py
beautifulsoup4==4.9.3
opencv-python-headless
scipy
statsmodels
sympy
datetime
googlesearch-python
researchpy
